Fix `$DRACULA_ARROW_ICON` getting parsed as a command-line flag to `print`
When `$DRACULA_ARROW_ICON` begins with an unescaped `-` character, it will be read as a command line flag by the `print` command. This causes the error `dracula_arrow:print:2: bad option: -> ` (for example) to get outputted on each new prompt line. This PR passes the `--` flag to `print` before passing `$DRACULA_ARROW_ICON` to ensure that `print` does not attempt to parse the arrow icon as a command-line flag.